Removed caching code + better search() behaviour
The old caching code could cause a query to be cached for ever,
returning always the same set of result for a specific query. This is
not a good behaviour, since search results change continuously. In the
future, this behaviour could be reintroduced with a better logic (for
example, check if the query was in cache longer than some hours or days)

Signed-off-by: Domenico Iezzi <domenico.iezzi.201@gmail.com>